refactor: move isCommitByUser and getFirstCommitDetails to their own respective api files
