+++ scripts/api/fetch-ongoing-workbench.js
@@ -1,11 +1,102 @@
 require('dotenv').config();
 const axios = require('axios');
 const { GITHUB_USERNAME, BASE_URL } = require('../config/config');
-const {
-  getLinkedIssueNumbers,
-  getPrActivityMeta,
-  getFirstCommitDetails,
-} = require('../utils/github-helpers');
+const { getLinkedIssueNumbers, getPrActivityMeta } = require('../utils/github-helpers');
+
+/**
+ * LOCAL HELPER: isCommitByUser (Ongoing Workbench Version)
+ * Strict logic to ignore web-flow and suggestions.
+ */
+function isCommitByUser(c, username) {
+  try {
+    const lowerUsername = username.toLowerCase();
+    const commitMessage = c.commit?.message || '';
+
+    // Ignore GitHub Web-flow actions and branch updates
+    if (c.committer?.login === 'web-flow') return false;
+    if (/suggestion|Merge branch|Merge remote-tracking|Merge pull request/i.test(commitMessage))
+      return false;
+
+    // Author check
+    const authorLogin = c.author?.login || '';
+    const authorEmail = c.commit?.author?.email?.toLowerCase() || '';
+    const authorName = c.commit?.author?.name?.toLowerCase() || '';
+
+    const isAuthorMatch =
+      authorLogin === username ||
+      authorEmail === `${lowerUsername}@users.noreply.github.com` ||
+      (authorEmail.endsWith('@users.noreply.github.com') &&
+        authorEmail.includes(`+${lowerUsername}@`)) ||
+      authorEmail.includes(lowerUsername) ||
+      authorName.includes(lowerUsername);
+
+    if (isAuthorMatch) return true;
+
+    // Co-authored check
+    if (
+      /Co-authored-by:/i.test(commitMessage) &&
+      commitMessage.toLowerCase().includes(lowerUsername)
+    ) {
+      return true;
+    }
+
+    return false;
+  } catch (e) {
+    return false;
+  }
+}
+
+/**
+ * LOCAL HELPER: getFirstCommitDetails (Ongoing Workbench Version)
+ */
+async function getFirstCommitDetails(
+  owner,
+  repo,
+  prNumber,
+  username,
+  commitCache,
+  axiosInstance,
+  prUpdatedAt = null
+) {
+  const prUrlKey = `/repos/${owner}/${repo}/pulls/${prNumber}`;
+  if (commitCache.has(prUrlKey)) {
+    const cached = commitCache.get(prUrlKey);
+    if (cached?.prUpdatedAt && prUpdatedAt && cached.prUpdatedAt === prUpdatedAt) return cached;
+  }
+
+  let result = null;
+  try {
+    let page = 1;
+    let allCommits = [];
+    while (true) {
+      const resp = await axiosInstance.get(`${prUrlKey}/commits?per_page=100&page=${page}`);
+      allCommits.push(...resp.data);
+      const linkHeader = resp.headers.link;
+      if (linkHeader && linkHeader.includes('rel="next"')) {
+        page++;
+        await new Promise((r) => setTimeout(r, 200));
+      } else {
+        break;
+      }
+    }
+
+    const userCommits = allCommits.filter((c) => isCommitByUser(c, username));
+    if (userCommits.length > 0) {
+      userCommits.sort((a, b) => new Date(a.commit.author.date) - new Date(b.commit.author.date));
+      result = {
+        firstCommitDate: userCommits[0].commit.author.date,
+        commitCount: userCommits.length,
+        prUpdatedAt,
+      };
+    } else {
+      result = { firstCommitDate: null, commitCount: 0, prUpdatedAt };
+    }
+  } catch (err) {
+    result = { firstCommitDate: null, commitCount: 0, prUpdatedAt };
+  }
+  commitCache.set(prUrlKey, result);
+  return result;
+}
 
 /**
  * FETCH ONGOING REVIEWS
